pyproject.toml
[tool.poetry]
name = "curso-flask"
version = "0.1.0"
description = "Projeto para estudo de Flask"
authors = ["Tiago Tardelli <tiagob.tardelli@gmail.com>"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
Flask-Login = "^0.6.2"
python-decouple = "^3.6"
Flask-WTF = "^1.0.1"
email-validator = "^1.3.0"
Bootstrap-Flask = "^2.1.0"
Flask-Mail = "^0.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
